mm_utils refactor
mrope fix
