+++ src/maxtext/multimodal/processor_qwen3_omni.py
@@ -1,4 +1,4 @@
-# Copyright 2023–2025 Google LLC
+# Copyright 2023–2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,6 +22,8 @@ import os
 from dataclasses import dataclass
 
 import numpy as np
+import jax
+import jax.numpy as jnp
 from PIL import Image
 
 try:
@@ -29,8 +31,7 @@ try:
 except ImportError:
   decord = None
 
-from MaxText.multimodal import utils as mm_utils
-from MaxText import multimodal_utils
+from maxtext.multimodal import utils as mm_utils
 from maxtext.utils import max_logging
 
 # Image constants.
@@ -55,6 +56,17 @@ SAMPLE_RATE = 16000  # Audio sampling rate in Hz.
 N_FFT = 400  # Number of FFT points for spectrogram computation.
 HOP_LENGTH = 160  # Number of samples between successive frames.
 DITHER = 0.0  # Amount of dithering to apply to audio signal.
+
+# Qwen3OmniMoe-specific processing
+QWEN3_OMNI_VISION_START_TOKEN = 151652  # <|vision_start|>
+QWEN3_OMNI_VISION_END_TOKEN = 151653  # <|vision_eos|>
+QWEN3_OMNI_IMAGE_TOKEN = 151655  # <|image_pad|>
+QWEN3_OMNI_VIDEO_TOKEN = 151656  # <|video_pad|>
+QWEN3_OMNI_AUDIO_START_TOKEN = 151669  # <|audio_start|>
+QWEN3_OMNI_AUDIO_END_TOKEN = 151648  # <|audio_eos|>
+QWEN3_OMNI_AUDIO_TOKEN = 151675  # <|audio_pad|>
+QWEN3_TEMPORAL_PATCH_SIZE = 2
+QWEN3_OMNI_IMAGE_SIZE = 768
 
 
 @dataclass
@@ -541,23 +553,19 @@ def add_extra_tokens_for_qwen3_omni(
     token = token_list[i]
 
     # Handle image tokens
-    if (
-        token == multimodal_utils.QWEN3_OMNI_IMAGE_TOKEN
-        and image_grid_thw is not None
-        and image_idx < len(image_grid_thw)
-    ):
+    if token == QWEN3_OMNI_IMAGE_TOKEN and image_grid_thw is not None and image_idx < len(image_grid_thw):
       grid = image_grid_thw[image_idx]
       num_image_tokens = int((grid[0] * grid[1] * grid[2]) // merge_length)
-      new_tokens.extend([multimodal_utils.QWEN3_OMNI_IMAGE_TOKEN] * num_image_tokens)
+      new_tokens.extend([QWEN3_OMNI_IMAGE_TOKEN] * num_image_tokens)
       image_idx += 1
 
     # Handle audio-in-video: <|vision_start|><|video_pad|><|vision_end|>
     elif (
         use_audio_in_video
-        and token == multimodal_utils.QWEN3_OMNI_VISION_START_TOKEN
+        and token == QWEN3_OMNI_VISION_START_TOKEN
         and i + 2 < len(token_list)
-        and token_list[i + 1] == multimodal_utils.QWEN3_OMNI_VIDEO_TOKEN
-        and token_list[i + 2] == multimodal_utils.QWEN3_OMNI_VISION_END_TOKEN
+        and token_list[i + 1] == QWEN3_OMNI_VIDEO_TOKEN
+        and token_list[i + 2] == QWEN3_OMNI_VISION_END_TOKEN
         and video_grid_thw is not None
         and video_idx < len(video_grid_thw)
     ):
@@ -579,52 +587,46 @@ def add_extra_tokens_for_qwen3_omni(
       video_token_indices = np.broadcast_to(video_token_indices, (num_frames, height, width)).flatten()
       video_token_indices = video_token_indices * second_per_grids[video_idx] * position_id_per_seconds
 
-      new_tokens.append(multimodal_utils.QWEN3_OMNI_VISION_START_TOKEN)
-      new_tokens.append(multimodal_utils.QWEN3_OMNI_AUDIO_START_TOKEN)
+      new_tokens.append(QWEN3_OMNI_VISION_START_TOKEN)
+      new_tokens.append(QWEN3_OMNI_AUDIO_START_TOKEN)
 
       video_data_idx = 0
       audio_data_idx = 0
 
       while video_data_idx < len(video_token_indices) and audio_data_idx < len(audio_token_indices):
         if video_token_indices[video_data_idx] <= audio_token_indices[audio_data_idx]:
-          new_tokens.append(multimodal_utils.QWEN3_OMNI_VIDEO_TOKEN)
+          new_tokens.append(QWEN3_OMNI_VIDEO_TOKEN)
           video_data_idx += 1
         else:
-          new_tokens.append(multimodal_utils.QWEN3_OMNI_AUDIO_TOKEN)
+          new_tokens.append(QWEN3_OMNI_AUDIO_TOKEN)
           audio_data_idx += 1
 
       while video_data_idx < len(video_token_indices):
-        new_tokens.append(multimodal_utils.QWEN3_OMNI_VIDEO_TOKEN)
+        new_tokens.append(QWEN3_OMNI_VIDEO_TOKEN)
         video_data_idx += 1
 
       while audio_data_idx < len(audio_token_indices):
-        new_tokens.append(multimodal_utils.QWEN3_OMNI_AUDIO_TOKEN)
+        new_tokens.append(QWEN3_OMNI_AUDIO_TOKEN)
         audio_data_idx += 1
 
-      new_tokens.append(multimodal_utils.QWEN3_OMNI_AUDIO_END_TOKEN)
-      new_tokens.append(multimodal_utils.QWEN3_OMNI_VISION_END_TOKEN)
+      new_tokens.append(QWEN3_OMNI_AUDIO_END_TOKEN)
+      new_tokens.append(QWEN3_OMNI_VISION_END_TOKEN)
 
       video_idx += 1
       audio_idx += 1
       i += 2
 
     # Handle video tokens (without audio-in-video)
-    elif (
-        token == multimodal_utils.QWEN3_OMNI_VIDEO_TOKEN
-        and video_grid_thw is not None
-        and video_idx < len(video_grid_thw)
-    ):
+    elif token == QWEN3_OMNI_VIDEO_TOKEN and video_grid_thw is not None and video_idx < len(video_grid_thw):
       grid = video_grid_thw[video_idx]
       num_video_tokens = int((grid[0] * grid[1] * grid[2]) // merge_length)
-      new_tokens.extend([multimodal_utils.QWEN3_OMNI_VIDEO_TOKEN] * num_video_tokens)
+      new_tokens.extend([QWEN3_OMNI_VIDEO_TOKEN] * num_video_tokens)
       video_idx += 1
 
     # Handle audio tokens (standalone, not in video)
-    elif (
-        token == multimodal_utils.QWEN3_OMNI_AUDIO_TOKEN and audio_lengths is not None and audio_idx < len(audio_lengths)
-    ):
+    elif token == QWEN3_OMNI_AUDIO_TOKEN and audio_lengths is not None and audio_idx < len(audio_lengths):
       num_audio_tokens = int(audio_lengths[audio_idx])
-      new_tokens.extend([multimodal_utils.QWEN3_OMNI_AUDIO_TOKEN] * num_audio_tokens)
+      new_tokens.extend([QWEN3_OMNI_AUDIO_TOKEN] * num_audio_tokens)
       audio_idx += 1
 
     # All other tokens pass through unchanged
@@ -634,3 +636,406 @@ def add_extra_tokens_for_qwen3_omni(
     i += 1
 
   return np.array(new_tokens, dtype=np.int32)
+
+
+def get_dummy_image_shape_for_init_qwen3_omni(batch_size):
+  """Return the shape of the dummy image for Qwen3-Omni model's initialization."""
+  image_shape = (
+      batch_size,
+      mm_utils.NUM_IMAGE_CHANNELS,
+      QWEN3_TEMPORAL_PATCH_SIZE,
+      QWEN3_OMNI_IMAGE_SIZE,  # image_size_for_vit (height)
+      QWEN3_OMNI_IMAGE_SIZE,  # video_num_frames
+  )
+  return image_shape
+
+
+def get_dummy_audio_shape_for_init_qwen3_omni(config):
+  """Return the shape of the dummy audio for Qwen3-Omni model's initialization."""
+  # Audio shape: (batch, num_mel_bins, audio_length)
+  # audio_length must be divisible by n_window * 2
+  chunk_size = config.n_window_for_audio * 2
+  audio_length = chunk_size * 4  # 4 chunks
+  audio_shape = (config.micro_batch_size_to_train_on, config.num_mel_bins_for_audio, audio_length)
+  return audio_shape
+
+
+# ==============================================================================
+# Qwen3-Omni Multimodal Position ID Utilities
+# ==============================================================================
+def _get_feat_extract_output_lengths(input_lengths: jax.Array) -> jax.Array:
+  """Computes the output length of the audio encoder convolutional layers.
+
+  The audio encoder processes audio in chunks of 100 samples, applying
+  multiple convolutional layers with stride 2. Each full 100-sample chunk
+  produces 13 output tokens, and remaining samples are processed separately.
+
+  Args:
+    input_lengths: Input audio sequence lengths. Shape: (batch,) or scalar.
+
+  Returns:
+    Output sequence lengths after audio encoding. Same shape as input.
+  """
+  input_lengths_leave = input_lengths % 100
+  feat_lengths = (input_lengths_leave - 1) // 2 + 1
+  output_lengths = ((feat_lengths - 1) // 2 + 1 - 1) // 2 + 1 + (input_lengths // 100) * 13
+  return output_lengths
+
+
+def get_llm_pos_ids_for_vision(
+    start_idx: int | jax.Array,
+    vision_idx: int,
+    spatial_merge_size: int,
+    t_index: jax.Array,
+    grid_hs: jax.Array,
+    grid_ws: jax.Array,
+) -> jax.Array:
+  """Computes 3D position IDs (temporal, height, width) for vision tokens.
+
+  Creates position embeddings for a grid of vision tokens representing an
+  image or video. For each temporal frame, generates a spatial grid of
+  (height, width) positions.
+
+  Args:
+    start_idx: Starting position ID value to add as offset.
+    vision_idx: Index of the current image/video being processed.
+    spatial_merge_size: Number of patches merged spatially (e.g., 2 means 2x2 patches → 1 token).
+    t_index: Temporal position for each frame. Shape: (num_frames,).
+    grid_hs: Height dimensions for all images/videos. Shape: (num_images,).
+    grid_ws: Width dimensions for all images/videos. Shape: (num_images,).
+
+  Returns:
+    3D position IDs with shape (3, num_vision_tokens) where:
+      - dim 0: temporal positions
+      - dim 1: height positions
+      - dim 2: width positions
+
+  Example:
+    If spatial_merge_size=2, grid_h=4, grid_w=4, num_frames=2:
+      - After merge: 2x2 grid per frame
+      - Total tokens: 2 frames x 2 x 2 = 8 tokens
+      - Output shape: (3, 8)
+      - t_index: [0, 0, 0, 0, 50, 50, 50, 50]
+      - h_index: [0, 0, 1, 1, 0, 0, 1, 1]
+      - w_index: [0, 1, 0, 1, 0, 1, 0, 1]
+  """
+  # Calculate effective spatial dimensions after merging patches
+  llm_grid_h = grid_hs[vision_idx] // spatial_merge_size
+  llm_grid_w = grid_ws[vision_idx] // spatial_merge_size
+
+  # Create height indices: [0, 0, ..., 0 (W times), 1, 1, ..., 1 (W times), ...]
+  # Shape: (num_frames, llm_grid_h, 1) -> expand -> (num_frames, llm_grid_h, llm_grid_w) -> flatten
+  h_index = jnp.arange(llm_grid_h).reshape(1, -1, 1).repeat(len(t_index), axis=0).repeat(llm_grid_w, axis=2).flatten()
+
+  # Create width indices: [0, 1, 2, ..., W-1, 0, 1, 2, ..., W-1, ...]
+  # Shape: (num_frames, 1, llm_grid_w) -> expand -> (num_frames, llm_grid_h, llm_grid_w) -> flatten
+  w_index = jnp.arange(llm_grid_w).reshape(1, 1, -1).repeat(len(t_index), axis=0).repeat(llm_grid_h, axis=1).flatten()
+
+  # Create temporal indices: [t0, t0, ..., t0 (HxW times), t1, t1, ..., t1 (HxW times), ...]
+  # Shape: (num_frames, 1) -> expand -> (num_frames, llm_grid_h * llm_grid_w) -> flatten
+  t_index_expanded = t_index.reshape(-1, 1).repeat(llm_grid_h * llm_grid_w, axis=1).flatten()
+
+  # Stack all three dimensions and add starting offset
+  _llm_pos_ids = jnp.stack([t_index_expanded, h_index, w_index])
+  llm_pos_ids = _llm_pos_ids + start_idx
+
+  return llm_pos_ids
+
+
+def get_chunked_index(token_indices: jax.Array, tokens_per_chunk: int, remove_index: int) -> list[tuple[int, int]]:
+  """Splits token index list into chunks based on token value ranges.
+
+  Given a list of monotonically increasing token indices, returns a list of
+  (start, end) index tuples representing slices where token values fall within
+  successive ranges of `tokens_per_chunk`.
+
+  Args:
+    token_indices: Monotonically increasing array of token index values. Shape: (seq_len,).
+    tokens_per_chunk: Chunk size threshold (e.g., 100 means first chunk has values < 100).
+    remove_index: Offset to subtract from token_indices before chunking.
+
+  Returns:
+    List of (start_idx, end_idx) tuples, each representing a chunk.
+
+  Example:
+    token_indices = [5, 10, 52, 105, 150, 250]
+    tokens_per_chunk = 100
+    remove_index = 0
+
+    Result: [(0, 3), (3, 5), (5, 6)]
+      - Chunk 0: indices 0-3 (values 5, 10, 52 are < 100)
+      - Chunk 1: indices 3-5 (values 105, 150 are >= 100 and < 200)
+      - Chunk 2: indices 5-6 (value 250 is >= 200)
+  """
+  chunks = []
+  i = 0
+  start_idx = 0
+  current_chunk = 1
+
+  while i < len(token_indices):
+    if token_indices[i] - remove_index >= current_chunk * tokens_per_chunk:
+      chunks.append((start_idx, i))
+      start_idx = i
+      current_chunk += 1
+    i += 1
+
+  # Append final chunk
+  chunks.append((start_idx, len(token_indices)))
+
+  return chunks
+
+
+def get_rope_index(
+    input_ids: np.ndarray,
+    image_grid_thw: np.ndarray | None = None,
+    video_grid_thw: np.ndarray | None = None,
+    attention_mask: np.ndarray | None = None,
+    use_audio_in_video: bool = False,
+    audio_lengths: np.ndarray | None = None,
+    second_per_grids: np.ndarray | None = None,
+    spatial_merge_size: int = 2,
+    position_id_per_seconds: int = 25,
+) -> tuple[np.ndarray, np.ndarray]:
+  """Calculate 3D RoPE position indices for multimodal sequences.
+
+  This function computes position IDs that encode both sequential (text) and
+  spatial-temporal (vision/audio) structure for Qwen3-Omni multimodal inputs.
+
+  For pure text sequences:
+    - All 3 dimensions receive the same sequential positions: [0, 1, 2, 3, 4]
+
+  For multimodal sequences with vision:
+    - Vision tokens get 3D positions (temporal, height, width)
+    - Text tokens continue sequentially from max(vision_pos) + 1
+    - Example with video (3 temporal patches, 2x2 spatial):
+        Vision temporal: [0, 0, 0, 0, 50, 50, 50, 50, 100, 100, 100, 100]
+        Vision height:   [0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1]
+        Vision width:    [0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1]
+        Text positions:  [101, 102, 103, 104, 105]
+
+  Args:
+    input_ids: Input token IDs. Shape: (batch, seq_len).
+    image_grid_thw: Image dimensions (temporal, height, width). Shape: (num_images, 3).
+    video_grid_thw: Video dimensions (temporal, height, width). Shape: (num_videos, 3).
+    attention_mask: Padding mask (1 = real token, 0 = padding). Shape: (batch, seq_len).
+    use_audio_in_video: If True, audio tokens are interleaved with video tokens.
+    audio_lengths: Audio sequence lengths. Shape: (num_audios,).
+    second_per_grids: Time interval per temporal grid (for videos). Shape: (num_videos,).
+    spatial_merge_size: Number of patches merged spatially (e.g., 2 for 2x2→1).
+    position_id_per_seconds: Temporal granularity (tokens per second, typically 25).
+
+  Returns:
+    A tuple of:
+      - position_ids: 3D position IDs. Shape: (3, batch, seq_len).
+      - mrope_position_deltas: Position offset for each sequence. Shape: (batch, 1).
+
+  Raises:
+    ValueError: If multimodal tokens are present but grid info is missing.
+  """
+  batch_size, seq_len = input_ids.shape
+
+  # Handle text-only case (no multimodal content)
+  if image_grid_thw is None and video_grid_thw is None:
+    if attention_mask is None:
+      attention_mask = np.ones_like(input_ids)
+
+    # Create sequential 1D positions
+    position_ids = np.cumsum(attention_mask.astype(np.float32), axis=-1) - 1
+    position_ids = np.where(attention_mask == 0, 1.0, position_ids)
+
+    # Expand to 3D (same value in all dimensions for text-only)
+    position_ids = np.broadcast_to(position_ids[np.newaxis, :, :], (3, batch_size, seq_len))
+
+    # Calculate deltas for each sequence
+    max_position_ids = np.max(position_ids, axis=(0, 2), keepdims=True).transpose(1, 0, 2)  # (batch, 1, 1)
+    mrope_position_deltas = max_position_ids.squeeze(-1) + 1 - np.sum(attention_mask, axis=-1, keepdims=True)
+
+    return position_ids, mrope_position_deltas
+
+  # Multimodal case: process each sequence in batch
+  if attention_mask is None:
+    attention_mask = np.ones_like(input_ids)
+
+  attention_mask_bool = attention_mask == 1
+  position_ids = np.zeros((3, batch_size, seq_len), dtype=jnp.float32)
+  mrope_position_deltas = []
+
+  # Process each sequence in the batch
+  for i in range(batch_size):
+    # Get valid tokens (non-padding)
+    valid_input_ids = input_ids[i][attention_mask_bool[i]]
+
+    # Count multimodal elements in this sequence
+    vision_start_indices = np.where(valid_input_ids == QWEN3_OMNI_VISION_START_TOKEN)[0]
+    vision_tokens = valid_input_ids[vision_start_indices + 1] if len(vision_start_indices) > 0 else np.array([])
+
+    audio_nums = np.sum(valid_input_ids == QWEN3_OMNI_AUDIO_START_TOKEN).item()
+    image_nums = np.sum(vision_tokens == QWEN3_OMNI_IMAGE_TOKEN).item() if len(vision_tokens) > 0 else 0
+    video_nums = (
+        (
+            np.sum(vision_tokens == QWEN3_OMNI_AUDIO_START_TOKEN).item()
+            if use_audio_in_video
+            else np.sum(vision_tokens == QWEN3_OMNI_VIDEO_TOKEN).item()
+        )
+        if len(vision_tokens) > 0
+        else 0
+    )
+
+    input_tokens = valid_input_ids.tolist()
+    llm_pos_ids_list = []
+    st = 0
+    remain_images = image_nums
+    remain_videos = video_nums
+    remain_audios = audio_nums
+    image_idx = 0
+    video_idx = 0
+    audio_idx = 0
+
+    multimodal_nums = image_nums + audio_nums if use_audio_in_video else image_nums + video_nums + audio_nums
+
+    # Process each multimodal element
+    for _ in range(multimodal_nums):
+      st_idx = llm_pos_ids_list[-1].max().item() + 1 if len(llm_pos_ids_list) > 0 else 0
+
+      # Find next vision or audio start token
+      if (QWEN3_OMNI_IMAGE_TOKEN in input_tokens or QWEN3_OMNI_VIDEO_TOKEN in input_tokens) and (
+          remain_videos > 0 or remain_images > 0
+      ):
+        try:
+          ed_vision_start = input_tokens.index(QWEN3_OMNI_VISION_START_TOKEN, st)
+        except ValueError:
+          ed_vision_start = len(input_tokens) + 1
+      else:
+        ed_vision_start = len(input_tokens) + 1
+
+      if QWEN3_OMNI_AUDIO_TOKEN in input_tokens and remain_audios > 0:
+        try:
+          ed_audio_start = input_tokens.index(QWEN3_OMNI_AUDIO_START_TOKEN, st)
+        except ValueError:
+          ed_audio_start = len(input_tokens) + 1
+      else:
+        ed_audio_start = len(input_tokens) + 1
+
+      min_ed = min(ed_vision_start, ed_audio_start)
+
+      # Add text tokens before multimodal element
+      text_len = min_ed - st
+      if text_len > 0:
+        text_pos = np.arange(text_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+        llm_pos_ids_list.append(text_pos)
+        st_idx += text_len
+
+      # Determine BOS/EOS token lengths
+      if min_ed == ed_vision_start and ed_vision_start + 1 == ed_audio_start:
+        bos_len, eos_len = 2, 2  # Audio in video
+      else:
+        bos_len, eos_len = 1, 1
+
+      # Add BOS token(s)
+      bos_pos = np.arange(bos_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+      llm_pos_ids_list.append(bos_pos)
+      st_idx += bos_len
+
+      # Process modality-specific content
+      # Audio Only
+      if min_ed == ed_audio_start:
+        audio_len = _get_feat_extract_output_lengths(audio_lengths[audio_idx]).item()
+        audio_pos = np.arange(audio_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+        llm_pos_ids_list.append(audio_pos)
+
+        st += int(text_len + bos_len + audio_len + eos_len)
+        audio_idx += 1
+        remain_audios -= 1
+
+      # Image Only
+      elif min_ed == ed_vision_start and input_tokens[ed_vision_start + 1] == QWEN3_OMNI_IMAGE_TOKEN:
+        grid_t = image_grid_thw[image_idx, 0].item()
+        grid_hs = image_grid_thw[:, 1]
+        grid_ws = image_grid_thw[:, 2]
+        t_index = np.arange(grid_t, dtype=np.float32) * 1 * position_id_per_seconds
+
+        image_pos = get_llm_pos_ids_for_vision(st_idx, image_idx, spatial_merge_size, t_index, grid_hs, grid_ws)
+        llm_pos_ids_list.append(image_pos)
+
+        image_len = int(np.prod(image_grid_thw[image_idx]).item() // (spatial_merge_size**2))
+        st += int(text_len + bos_len + image_len + eos_len)
+        image_idx += 1
+        remain_images -= 1
+
+      # Video Only
+      elif min_ed == ed_vision_start and input_tokens[ed_vision_start + 1] == QWEN3_OMNI_VIDEO_TOKEN:
+        grid_t = video_grid_thw[video_idx, 0].item()
+        grid_hs = video_grid_thw[:, 1]
+        grid_ws = video_grid_thw[:, 2]
+        t_index = np.arange(grid_t, dtype=np.float32) * second_per_grids[video_idx].item() * position_id_per_seconds
+
+        video_pos = get_llm_pos_ids_for_vision(st_idx, video_idx, spatial_merge_size, t_index, grid_hs, grid_ws)
+        llm_pos_ids_list.append(video_pos)
+
+        video_len = int(np.prod(video_grid_thw[video_idx]).item() // (spatial_merge_size**2))
+        st += int(text_len + bos_len + video_len + eos_len)
+        video_idx += 1
+        remain_videos -= 1
+
+      # Audio in Video (interleaved)
+      elif min_ed == ed_vision_start and ed_vision_start + 1 == ed_audio_start:
+        audio_len = _get_feat_extract_output_lengths(audio_lengths[audio_idx]).item()
+        audio_llm_pos_ids = np.arange(audio_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+
+        grid_t = video_grid_thw[video_idx, 0].item()
+        grid_hs = video_grid_thw[:, 1]
+        grid_ws = video_grid_thw[:, 2]
+        t_index = np.arange(grid_t, dtype=np.float32) * second_per_grids[video_idx].item() * position_id_per_seconds
+
+        video_llm_pos_ids = get_llm_pos_ids_for_vision(st_idx, video_idx, spatial_merge_size, t_index, grid_hs, grid_ws)
+
+        # Interleave audio and video based on temporal ordering
+        video_data_index = 0
+        audio_data_index = 0
+        while video_data_index < video_llm_pos_ids.shape[1] and audio_data_index < audio_llm_pos_ids.shape[1]:
+          if video_llm_pos_ids[0, video_data_index] <= audio_llm_pos_ids[0, audio_data_index]:
+            llm_pos_ids_list.append(video_llm_pos_ids[:, video_data_index : video_data_index + 1])
+            video_data_index += 1
+          else:
+            llm_pos_ids_list.append(audio_llm_pos_ids[:, audio_data_index : audio_data_index + 1])
+            audio_data_index += 1
+
+        # Append remaining tokens
+        if video_data_index < video_llm_pos_ids.shape[1]:
+          llm_pos_ids_list.append(video_llm_pos_ids[:, video_data_index:])
+        if audio_data_index < audio_llm_pos_ids.shape[1]:
+          llm_pos_ids_list.append(audio_llm_pos_ids[:, audio_data_index:])
+
+        video_len = int(np.prod(video_grid_thw[video_idx]).item() // (spatial_merge_size**2))
+        st += int(text_len + bos_len + audio_len + video_len + eos_len)
+
+        audio_idx += 1
+        video_idx += 1
+        remain_videos -= 1
+        remain_audios -= 1
+
+      # Add EOS token(s)
+      st_idx = llm_pos_ids_list[-1].max().item() + 1 if len(llm_pos_ids_list) > 0 else 0
+      eos_pos = np.arange(eos_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+      llm_pos_ids_list.append(eos_pos)
+
+    # Add any remaining text tokens
+    if st < len(input_tokens):
+      st_idx = llm_pos_ids_list[-1].max().item() + 1 if len(llm_pos_ids_list) > 0 else 0
+      text_len = len(input_tokens) - st
+      remaining_text_pos = np.arange(text_len).reshape(1, -1).repeat(3, axis=0) + st_idx
+      llm_pos_ids_list.append(remaining_text_pos)
+
+    # Concatenate all position IDs for this sequence
+    llm_positions = np.concatenate(llm_pos_ids_list, axis=1)
+
+    # Place into position_ids tensor at valid positions
+    valid_positions = np.where(attention_mask_bool[i])[0]
+    position_ids[:, i, valid_positions] = llm_positions
+
+    # Calculate delta for this sequence
+    mrope_position_deltas.append(llm_positions.max().item() + 1 - len(valid_input_ids))
+
+  mrope_position_deltas = np.array(mrope_position_deltas).reshape(batch_size, 1)
+
+  return position_ids, mrope_position_deltas
