+++ src/maxtext/multimodal/utils.py
@@ -1,4 +1,4 @@
-# Copyright 2023–2025 Google LLC
+# Copyright 2023–2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,9 +18,17 @@ import os
 from dataclasses import dataclass
 from typing import Optional, Union
 
-import librosa
 import numpy as np
+import jax
+import jax.numpy as jnp
 from PIL import Image
+
+try:
+  import librosa  # pytype: disable=import-error
+except ImportError:
+  librosa = None
+
+NUM_IMAGE_CHANNELS = 3  # RGB
 
 
 @dataclass
@@ -52,6 +60,13 @@ class PreprocessorOutput:
   audio_mask: None | np.ndarray = None
 
 
+def convert_to_RGB(image):
+  """Convert image to RGB format."""
+  if image.mode != "RGB":
+    image = image.convert("RGB")
+  return image
+
+
 def load_image_from_path(image_path):
   """Loads an image from a given file path and returns a np.array."""
   if not os.path.isfile(image_path):
@@ -80,6 +95,112 @@ def normalize_images(images, mean, std):
   images -= np.asarray(mean)
   images /= np.asarray(std)
   return images
+
+
+def merge_mm_embeddings(
+    text_embeddings: np.ndarray | jnp.ndarray,
+    multimodal_embeddings: np.ndarray | jnp.ndarray,
+    mask,
+    token_masks: np.ndarray | jnp.ndarray | None = None,
+) -> np.ndarray | jnp.ndarray:
+  """Merges text and multimodal (vision/audio) embeddings based on a mask.
+
+  This function handles two primary formats for multimodal embeddings:
+  1. Tiled Format (e.g., Llama4 vision): Embeddings are provided as a batch of
+     images and their tiles, with shape (B * N, T, K, D). These are flattened
+     into a single sequence of tokens per batch item.
+  2. Simple Format (e.g., Gemma3 vision, Qwen3-Omni audio): Embeddings are provided as
+     (B, N, K, D) and are flattened into a sequence of tokens.
+
+  Args:
+    text_embeddings: (B, S, D) array of text embeddings.
+    multimodal_embeddings: Multimodal embeddings (vision/audio) in one of two formats:
+      - (B * N, T, K, D) for tiled inputs.
+      - (B, N, K, D) for simple inputs.
+      (B=batch_size, S=seq_len, D=embedding_dim, N=num_images/audio_chunks,
+       T=num_tiles, K=toks_per_image/audio_chunk)
+    mask: (B, S) boolean or integer array where non-zero positions
+      indicate where multimodal embeddings should be placed.
+    token_masks: (Optional) A mask for the multimodal tokens.
+      - (B * N, T) for tiled inputs, indicating valid tiles.
+      - If None, all multimodal embeddings are assumed to be valid.
+
+  Returns:
+    A (B, S, D) array of merged embeddings.
+  """
+  # Input Validation and Shape Unpacking
+  batch_size, _, d_model = text_embeddings.shape
+  # The number of tokens per image/tile/audio_chunk is the second to last dimension.
+  num_toks_per_token = multimodal_embeddings.shape[-2]
+
+  if d_model != multimodal_embeddings.shape[-1]:
+    raise ValueError(
+        "Embedding dimension mismatch between text and multimodal embeddings:"
+        f" {d_model} vs {multimodal_embeddings.shape[-1]}"
+    )
+
+  # Reshape Multimodal Embeddings to a unified (B, S_mm, D) format
+  # This single reshape robustly handles both documented cases:
+  # Case 1: (B * N, T, K, D) -> (B, N*T*K, D)
+  # Case 2: (B, N, K, D) -> (B, N*K, D)
+  flat_multimodal_embeddings = multimodal_embeddings.reshape(batch_size, -1, d_model)
+
+  # Process Optional Token Masks
+  flat_token_masks_processed = None
+  if token_masks is not None:
+    # Handle the tiled case where token_masks batch dimension is (B * N)
+    if token_masks.shape[0] != batch_size:
+      if token_masks.shape[0] % batch_size != 0:
+        raise ValueError(
+            "Batch dimension of token_masks must be a multiple of the text"
+            f" batch size. Got {token_masks.shape[0]} and {batch_size}."
+        )
+      # Reshape from (B * N, T) to (B, N * T)
+      flat_tile_masks = token_masks.reshape(batch_size, -1)
+    else:
+      # This handles cases where token_masks is already (B, ...)
+      flat_tile_masks = token_masks.reshape(batch_size, -1)
+
+    # Expand the tile-level mask to a token-level mask to match the embeddings.
+    # A mask of shape (B, N*T) becomes (B, N*T*K) by repeating each element K times.
+    flat_token_masks_processed = jnp.repeat(flat_tile_masks, repeats=num_toks_per_token, axis=1)
+
+  # Vmap the inner merge function over the batch dimension
+  return jax.vmap(
+      _merge_mm_embeddings_inner,  # Assumes this function is defined elsewhere
+      in_axes=(0, 0, 0, None if flat_token_masks_processed is None else 0),
+  )(text_embeddings, flat_multimodal_embeddings, mask, flat_token_masks_processed)
+
+
+def _merge_mm_embeddings_inner(
+    text_embeddings: jnp.ndarray,
+    multimodal_embeddings: jnp.ndarray,
+    mask: jnp.ndarray | None = None,
+    token_mask: jnp.ndarray | None = None,
+) -> jnp.ndarray:
+  """`merge_mm_embeddings` without batch dimension."""
+
+  if token_mask is not None:
+    # This logic packs valid multimodal tokens to the front of the array.
+    # It correctly handles cases where some multimodal tokens are just padding.
+    sort_indices = jnp.argsort(-token_mask)  # Sorts descending, putting 1s first
+    multimodal_embeddings = multimodal_embeddings[sort_indices]
+
+  # Find positions in the text sequence to place the multimodal embeddings.
+  # The `size` argument ensures a fixed shape for JIT compilation.
+  target_pos = jnp.nonzero(mask, size=multimodal_embeddings.shape[0])
+  target_pos = target_pos[0]  # jnp.nonzero returns a tuple of arrays
+
+  # Save the embedding at the first position.
+  first_pos_embedding = text_embeddings[0]
+
+  # Perform the insertion.
+  merged = text_embeddings.at[target_pos, :].set(multimodal_embeddings)
+
+  # Restore the first position's embedding, in case it was overwritten.
+  merged = merged.at[0].set(first_pos_embedding)
+
+  return merged
 
 
 # Following audio functions derived from the HuggingFace implementation
@@ -661,7 +782,8 @@ def load_audio(data_path: str, sample_rate: int = 16000) -> np.ndarray:
   """
   if not os.path.isfile(data_path):
     raise FileNotFoundError(f"Audio file not found at path {data_path}. Please specify a valid audio file path")
-
+  if librosa is None:
+    raise ImportError("librosa is required for audio processing but not installed.")
   try:
     audio = librosa.load(data_path, sr=sample_rate)[0]
     return audio
